+++ src/integrators/rk4.cpp
@@ -14,8 +14,8 @@ namespace rk4 {
     /// @param star_mass Mass of the Star
     /// @param G Gravitational constant
     /// @return Returns acceleration of the orbiting object at the given position
-    Vec2 acceleration_update(const Vec2& star_pos, const Vec2& planet_pos, const float star_mass, const float G) {
-        float distance = physics::find_distance(star_pos, planet_pos);
+    Vec2 acceleration_update(const Vec2& star_pos, const Vec2& planet_pos, const double star_mass, const double G) {
+        double distance = physics::find_distance(star_pos, planet_pos);
 
         return physics::find_g_acceleration(star_mass, distance, star_pos, planet_pos, G);
     }
@@ -29,7 +29,7 @@ namespace rk4 {
     /// @param G Gravitational constant
     /// @return Returns the updated position value using the fourth-order Runge-Kutta method
     Vec2 position_update(const Vec2& cur_position, const Vec2& cur_velocity,
-                         const Vec2& star_pos, const float star_mass, const float t, const float G) {
+                         const Vec2& star_pos, const double star_mass, const double t, const double G) {
         Vec2 k1_position = cur_velocity;
         Vec2 k1_velocity = acceleration_update(star_pos, cur_position, star_mass, G);
 
@@ -53,7 +53,7 @@ namespace rk4 {
     /// @param G Gravitational constant
     /// @return Returns the updated velocity value using the fourth-order Runge-Kutta method
     Vec2 velocity_update(const Vec2& cur_position, const Vec2& cur_velocity,
-                         const Vec2& star_pos, const float star_mass, const float t, const float G) {
+                         const Vec2& star_pos, const double star_mass, const double t, const double G) {
         Vec2 k1_position = cur_velocity;
         Vec2 k1_velocity = acceleration_update(star_pos, cur_position, star_mass, G);
 
@@ -77,17 +77,17 @@ namespace rk4 {
     /// @param iterations Number of iterations algorithm should run
     /// @param G Gravitational constant
     integrators::RunSummary run_rk4(const Vec2& star_pos, const Vec2& planet_pos,
-                                    const float star_mass, const float planet_mass,
-                                    const float t, const int iterations, const float G,
+                                    const double star_mass, const double planet_mass,
+                                    const double t, const int iterations, const double G,
                                     const std::string& output_csv) {
         // Find initial conditions of the problem
-        float distance = physics::find_distance(star_pos, planet_pos);
-        float orbital_speed = physics::find_velocity(star_mass, distance, G);
+        double distance = physics::find_distance(star_pos, planet_pos);
+        double orbital_speed = physics::find_velocity(star_mass, distance, G);
         Vec2 displacement = planet_pos;
         Vec2 velocity = physics::find_vel_direction(star_pos, planet_pos, distance ,orbital_speed);
-        float initial_energy = physics::find_energy_conservation(star_mass, planet_mass, G,
+        double initial_energy = physics::find_energy_conservation(star_mass, planet_mass, G,
                                                                     distance, velocity);
-        float initial_angular_momentum = physics::find_angular_momentum(planet_mass, velocity, planet_pos - star_pos);
+        double initial_angular_momentum = physics::find_angular_momentum(planet_mass, velocity, planet_pos - star_pos);
         integrators::RunRegistry registry(output_csv);
         registry.record(0, 0.0, displacement, velocity, distance, initial_energy, initial_angular_momentum, 0.0, 0.0);
 
@@ -109,14 +109,14 @@ namespace rk4 {
             // Calculate conservations and errors
 
             // Energy
-            float energy_conservation = physics::find_energy_conservation(star_mass, planet_mass, G,
+            double energy_conservation = physics::find_energy_conservation(star_mass, planet_mass, G,
                                                                           distance, velocity);
-            float energy_error = physics::energy_error(energy_conservation, initial_energy);
+            double energy_error = physics::energy_error(energy_conservation, initial_energy);
 
             // Angular Momentum
-            float angular_momentum_conservation = physics::find_angular_momentum(planet_mass, velocity,
+            double angular_momentum_conservation = physics::find_angular_momentum(planet_mass, velocity,
                                                                                  displacement - star_pos);
-            float angular_error = physics::angular_error(angular_momentum_conservation, initial_angular_momentum);
+            double angular_error = physics::angular_error(angular_momentum_conservation, initial_angular_momentum);
             registry.record(i, i * t, displacement, velocity, distance, energy_conservation,
                             angular_momentum_conservation, energy_error, angular_error);
         }
