+++ src/physics.cpp
@@ -13,8 +13,8 @@ namespace physics {
     /// @param b The lighter orbiting body.
     /// @details Follows a simple Pythagorean of squaring two distances and subtracting them.
     ///          Two subtracted squares are added and their square root is taken.
-    /// @return Returns the equation as a float directly.
-    float find_distance(const Vec2& a, const Vec2& b) {
+    /// @return Returns the equation as a double directly.
+    double find_distance(const Vec2& a, const Vec2& b) {
         return sqrt(pow(a.x - b.x, 2) + pow(a.y - b.y, 2));
     }
 
@@ -27,9 +27,9 @@ namespace physics {
     /// @details Simplifies Newton's `F=ma` and `F=G*(Mm/r^3)*r_vector.
     ///          Finds vector distance between the two objects for r_vector.
     /// @return Returns the acceleration
-    Vec2 find_g_acceleration(float m_a, float distance, const Vec2& a, const Vec2& b, const float G) {
-        // Narrowing conversion that auto-promotes int to float to calculate the gravitational component
-        float grav_component = (-G * m_a) / pow(distance, 3);
+    Vec2 find_g_acceleration(double m_a, double distance, const Vec2& a, const Vec2& b, const double G) {
+        // Narrowing conversion that auto-promotes int to double to calculate the gravitational component
+        double grav_component = (-G * m_a) / pow(distance, 3);
         Vec2 acceleration = grav_component * (b - a);
 
         return acceleration;
@@ -40,7 +40,7 @@ namespace physics {
     /// @param distance Distance between two object
     /// @param G Constant of gravitational attraction
     /// @return Returns the circular orbital velocity of the object
-    float find_velocity(float m, float distance, const float G) {
+    double find_velocity(double m, double distance, const double G) {
         return sqrt((G*m) / distance);
     }
 
@@ -50,7 +50,7 @@ namespace physics {
     /// @param distance Distance between the two objects.
     /// @param vel Current velocity of the orbiting object.
     /// @return Returns vector velocity
-    Vec2 find_vel_direction(const Vec2& a, const Vec2& b, float distance, float vel) {
+    Vec2 find_vel_direction(const Vec2& a, const Vec2& b, double distance, double vel) {
         Vec2 radius_vector = b - a;
         Vec2 unit_perpendicular = radius_vector.perpendicular() / distance;
         Vec2 vector_vel = vel * unit_perpendicular;
@@ -65,7 +65,7 @@ namespace physics {
     /// @param distance Distance between the two objects
     /// @param b_velocity Orbiting Object's Velocity
     /// @return Returns the energy conservation of the integrator based on the Hamiltonian framework.
-    float find_energy_conservation(float M, float m, const float G, float distance, const Vec2& b_velocity) {
+    double find_energy_conservation(double M, double m, const double G, double distance, const Vec2& b_velocity) {
         return 0.5 * m * (pow(b_velocity.x, 2) + pow(b_velocity.y, 2)) - (G * M * m)/distance;
     }
 
@@ -73,7 +73,7 @@ namespace physics {
     /// @param e_n Energy of the current timestep
     /// @param e_0 Energy at the start
     /// @return Returns the relative error of the energy system
-    float energy_error(float e_n, float e_0) {
+    double energy_error(double e_n, double e_0) {
         return std::abs(e_n - e_0) / std::abs(e_0);
     }
 
@@ -82,7 +82,7 @@ namespace physics {
     /// @param b_velocity Velocity of the orbiting object
     /// @param b Position of the orbiting object
     /// @return Returns current angular momentum
-    float find_angular_momentum(float m, const Vec2& b_velocity, const Vec2&b) {
+    double find_angular_momentum(double m, const Vec2& b_velocity, const Vec2&b) {
         return m * ((b.x * b_velocity.y) - (b.y * b_velocity.x));
     }
 
@@ -90,7 +90,7 @@ namespace physics {
     /// @param l_n Angular momentum of the current timestep
     /// @param l_0 Angular momentum at the start
     /// @return Returns the relative error of the angular momentum
-    float angular_error(float l_n, float l_0) {
+    double angular_error(double l_n, double l_0) {
         return std::abs(l_n - l_0) / std::abs(l_0);
     }
 }
